requirement.txt
torch
torchvision
pyyaml
tqdm
numpy
scipy
matplotlib
pillow
timm==0.5.4
